+++ pages/courses/[dept]/[num].jsx
@@ -3,12 +3,12 @@ import React, { useEffect } from 'react';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
 import { useDispatch, useSelector } from 'react-redux';
-import CourseInfoSubtitle from '../../../components/CourseInfoSubtitle';
-import Glance from '../../../components/Glance';
-import CourseInfoTitle from '../../../components/CourseInfoTitle';
-import Offered from '../../../components/Offered';
-import Medians, { convertMedian } from '../../../components/Medians';
-import StudentsSay from '../../../components/StudentsSay';
+import CourseInfoSubtitle from '../../../components/courseInfo/CourseInfoSubtitle';
+import Glance from '../../../components/courseInfo/Glance';
+import CourseInfoTitle from '../../../components/courseInfo/CourseInfoTitle';
+import Offered from '../../../components/courseInfo/Offered';
+import Medians, { convertMedian } from '../../../components/courseInfo/Medians';
+import StudentsSay from '../../../components/courseInfo/StudentsSay';
 import { fetchCourse } from '../../../actions';
 import getPrereqs from '../../../data/courseinfohelpers';
 import styles from '../../../styles/ExploreHome.module.css';
